Terapkan revisi titik 15.54, pulihkan nilai melipat, dan hapus UID-2391
Berkas revisi membawa perbaikan UID-0525 dan UID-0559: tipe simpang berubah
dari bersinyal menjadi tak bersinyal dan medianya dari Videotron menjadi Neon
Box, sehingga sewa NS A dan NS C ikut turun.

Perbaikan UID-2420 tidak terbawa pada berkas revisi, nilainya kembali ke
keadaan sebelum perbaikan. Delapan kolomnya dikembalikan dari data terakhir.

Tiga belas nilai rupiah terbaca negatif pada berkas revisi karena kolomnya
bertipe MEDIUMINT yang hanya 4 byte, sehingga nilai di atas 2,1 miliar melipat
saat berkas dibuka-tulis ulang oleh QGIS. Semuanya persis nilai benar dikurangi
2^32 dan telah dipulihkan. Penyebabnya ada pada tools/ekspor_gpkg.py yang
menuliskan kolom bilangan bulat sebagai MEDIUMINT; kini memakai INTEGER yang
pada GeoPackage berarti 8 byte, dan terbaca sebagai Integer64 oleh GDAL.

UID-2391 (PWK-LBJ-BUN-SIM-002) dihapus beserta empat titik POV miliknya. Titik
ini duplikat dari UID-2396 yang dipertahankan: berjarak 39 m, sama-sama Simpang
Bersinyal di Jl. Pembangunan, dan nilai tarifnya identik.

Hasil: 2.481 titik reklame dan 906 titik POV. Tidak ada lagi nilai rupiah
negatif, tidak ada ID_TITIK ganda.

+++ tools/ekspor_gpkg.py
@@ -26,7 +26,10 @@ WKT_4326 = (
     'UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],'
     'AUTHORITY["EPSG","4326"]]'
 )
-TIPE_SQL = {"integer": "MEDIUMINT", "number": "DOUBLE", "text": "TEXT"}
+# INTEGER pada GeoPackage berarti 8 byte. MEDIUMINT hanya 4 byte, sehingga
+# nilai rupiah di atas 2,1 miliar akan melipat jadi negatif saat berkas
+# dibuka-tulis ulang oleh QGIS.
+TIPE_SQL = {"integer": "INTEGER", "number": "DOUBLE", "text": "TEXT"}
 
 
 def blob_point(lon, lat, srs=4326):
